Moving SFM poller to library function, fixing overnight timers
