+++ CHANGES.md
@@ -1,13 +1,5 @@
 # List of changes to be made in the 2.0 release
 - look for GraphQL vehicle API
-- change strike to server load, load everything in parent load, reuse in component, remove API
-- consider moving all pollers to separate library directory:
-   - stop: DONE
-   - route: DONE
-   - strikes: TODO
-   - sfm: TODO
-   - images: NOT PLANNED
-   - searches: NOT PLANNED  (see GET search)
 - investigate possibility of using enhanced forms with method="GET" with search queries
   - if possible: move search to library as poller, call with form
 - add MATO as source in about section
