+++ src/routes/sfm/[stop]/+layout.server.ts
@@ -1,7 +1,7 @@
 import type { stopDB } from '$lib/stopDB';
 import type { LayoutServerLoad } from './$types';
 import type { Collection } from "mongodb";
-import { poll } from './poll';
+import { poll } from '$lib/poll/sfm';
 
 //DB data is returned first, passed to the client load page. This never changes, so it shouldn't expire
 export const load: LayoutServerLoad = async ({ params, locals, depends }) => {
